refactor: reimplementado a forma de criar order na aplicação

+++ src/main/java/br/selene/projectseleneback/domain/order/dto/CreateTicketDTO.java
@@ -1,8 +1,8 @@
 package br.selene.projectseleneback.domain.order.dto;
 
 public record CreateTicketDTO(
-        int eventId,
-        int categoryId,
+        Long eventId,
+        Long categoryId,
         int quantity
 ){
