+++ src/main/java/br/selene/projectseleneback/domain/order/dto/RequestCreateOrderDTO.java
@@ -3,7 +3,7 @@ package br.selene.projectseleneback.domain.order.dto;
 import java.util.List;
 
 public record RequestCreateOrderDTO(
-        int customerId,
+        Long customerId,
         List<CreateTicketDTO> tickets
 ) {}
