+++ src/main/java/br/selene/projectseleneback/domain/ticketCategory/dto/CreateTicketCategoryDTO.java
@@ -1,4 +1,4 @@
 package br.selene.projectseleneback.domain.ticketCategory.dto;
 
-public record CreateTicketCategoryDTO (int price, String description, int quantity){
+public record CreateTicketCategoryDTO (Long price, String description, int quantity){
 }
